Moved DSL classes to separate files
