+++ lib/tlaw/dsl/api_wrapper.rb
@@ -1,0 +1,11 @@
+require_relative 'namespace_wrapper'
+
+module TLAW
+  module DSL
+    class APIWrapper < NamespaceWrapper
+      def base(url)
+        @object.base_url = url
+      end
+    end
+  end
+end
