+++ lib/tlaw/dsl/endpoint_wrapper.rb
@@ -1,0 +1,8 @@
+require_relative 'base_wrapper'
+
+module TLAW
+  module DSL
+    class EndpointWrapper < BaseWrapper
+    end
+  end
+end
